Site estilizado!

+++ modulo2/projeto-labefy/src/pages/ManagePlaylists/ManagePlaylists.js
@@ -2,7 +2,8 @@ import React from "react"
 import axios from "axios"
 import { BASE_URL } from "../../constants/urls"
 import { axiosConfig } from "../../constants/urls"
-import { PlaylistContainer } from "./styled"
+import { PlaylistContainer, BodyContainer, FormContainer, DescriptionContainer } from "./styled"
+import Trashcan from "../../images/trashcan.svg"
 
 export default class ManagePlaylists extends React.Component {
     state = {
@@ -10,12 +11,12 @@ export default class ManagePlaylists extends React.Component {
         playlists: []
     }
 
-    componentDidMount () {
+    componentDidMount() {
         this.getAllPlaylists()
     }
 
     onChangePlaylistName = (event) => {
-        this.setState({playlistNameInput: event.target.value})
+        this.setState({ playlistNameInput: event.target.value })
     }
 
     createPlaylist = () => {
@@ -24,58 +25,66 @@ export default class ManagePlaylists extends React.Component {
         }
 
         axios.post(BASE_URL, body, axiosConfig)
-        .then(() => {
-            alert("Playlist criada com sucesso!")
-            this.setState({playlistNameInput: ""})
-            this.getAllPlaylists()
-        })
-        .catch((error) => {
-            alert(error.response.data.message)
-        })
+            .then(() => {
+                alert("Playlist criada com sucesso!")
+                this.setState({ playlistNameInput: "" })
+                this.getAllPlaylists()
+            })
+            .catch((error) => {
+                alert(error.response.data.message)
+            })
     }
 
     getAllPlaylists = () => {
         axios.get(BASE_URL, axiosConfig)
-        .then((response) => {
-            this.setState({playlists: response.data.result.list})
-        })
-        .catch((error) => {
-            alert(error.response.data)
-        })
+            .then((response) => {
+                this.setState({ playlists: response.data.result.list })
+            })
+            .catch((error) => {
+                alert(error.response.data)
+            })
     }
 
     deletePlaylist = (id) => {
-        axios.delete(`${BASE_URL}/${id}`, axiosConfig)
-        .then(() => {
-            alert("Playlist deletada com sucesso!")
-            this.getAllPlaylists()
-        })
-        .catch((error) => {
-            alert(error.response.data)
-        })
+        if (window.confirm("Tem certeza que deseja deletar essa playlist?")) {
+            axios.delete(`${BASE_URL}/${id}`, axiosConfig)
+                .then(() => {
+                    alert("Playlist deletada com sucesso!")
+                    this.getAllPlaylists()
+                })
+                .catch((error) => {
+                    alert(error.response.data)
+                })
+        }
     }
 
-    render () {
-        const listOfPlaylists = this.state.playlists.map ((playlist) => {
+    render() {
+        const listOfPlaylists = this.state.playlists.map((playlist) => {
             return <PlaylistContainer key={playlist.id}>
-                <p onClick={() => this.props.switchToTracks(playlist)}>{playlist.name}</p>
-                <button onClick={() => this.deletePlaylist(playlist.id)}>Deletar Playlist</button>
+                <p>{this.state.playlists.indexOf(playlist) + 1}</p>
+                <p id="playlistName" onClick={() => this.props.switchToTracks(playlist)}>{playlist.name}</p>
+                <button id="deleteButton" onClick={() => this.deletePlaylist(playlist.id)}><img src={Trashcan} /></button>
             </PlaylistContainer>
         })
 
-        return (<>
-            <div>
-                <input 
-                    type="text" 
-                    value={this.state.playlistNameInput} 
+        return (<BodyContainer>
+            <h2>Playlists</h2>
+            <FormContainer>
+                <input
+                    type="text"
+                    value={this.state.playlistNameInput}
                     onChange={this.onChangePlaylistName}
                     placeholder="Nome da sua Playlist"
                 />
                 <button onClick={this.createPlaylist}>Criar Playlist</button>
-            </div>
+            </FormContainer>
+            <DescriptionContainer>
+                <p id="hash">#</p>
+                <p id="title">Título</p>
+            </DescriptionContainer>
             <div>
-                {listOfPlaylists}               
+                {listOfPlaylists}
             </div>
-        </>)
+        </BodyContainer>)
     }
 }
